+++ NexusKrop.IceCube/IO/NBinaryWriter.cs
@@ -13,12 +13,8 @@
 // limitations under the License.
 
 namespace NexusKrop.IceCube.IO;
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
 /// <summary>
 /// A <see cref="BinaryWriter"/> that implements <see cref="IBinaryWriter"/>.