+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0-alpha] - 2023/3/3
+
 ### Added
 
 - Added Key-to-Value Container that can be formatted as binary, and stores primitive data
@@ -75,7 +77,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://codeberg.org/NexusKrop/IceCube/compare/v0.1.4-alpha...HEAD
+[Unreleased]: https://codeberg.org/NexusKrop/IceCube/compare/v0.2.0-alpha...HEAD
+[0.2.0-alpha]: https://codeberg.org/NexusKrop/IceCube/compare/v0.1.4-alpha...v0.2.0-alpha
 [0.1.4-alpha]: https://codeberg.org/NexusKrop/IceCube/compare/v0.1.3-alpha...v0.1.4-alpha
 [0.1.3-alpha]: https://codeberg.org/NexusKrop/IceCube/compare/v0.1.2-alpha...v0.1.3-alpha
 [0.1.2-alpha]: https://codeberg.org/NexusKrop/IceCube/compare/v0.1.1-alpha...v0.1.2-alpha